feat: move hidden icons to library header
